requirements.txt
bs4
flask
flask-cors
flask-login
flask-sqlalchemy
folium
google-genai
python-dotenv
redis
regex
requests
werkzeug

